# resources/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arkon API",
        "version": "1.0.0",
        "description": "Arkon's public developer API. Guide: docs/api.md. Anonymous requests read published content; a personal access token (Authorization: Bearer) adds the scopes it was given, never more than its user's role allows. Errors: {\"error\": {\"code\", \"message\", \"fields\"?}}."
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "The site the API is reached at"
        }
    ],
    "security": [
        [],
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/forms": {
            "get": {
                "operationId": "listForms",
                "tags": [
                    "Forms"
                ],
                "summary": "All forms with status (read:forms)",
                "security": [
                    {
                        "bearerAuth": [
                            "read:forms"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "description": "Name.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "description": "Default all.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "active",
                                "draft",
                                "inactive",
                                "trash"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "Page number, from 1.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data",
                                        "meta",
                                        "links"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FormSummary"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
        "/forms/{id}": {
            "get": {
                "operationId": "getForm",
                "tags": [
                    "Forms"
                ],
                "summary": "The definition of a form a live page uses",
                "security": [
                    [],
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Form"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/forms/{id}/submissions": {
            "post": {
                "operationId": "submitForm",
                "tags": [
                    "Forms"
                ],
                "summary": "Send a submission (validated, stored encrypted, rate limited)",
                "security": [
                    [],
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": false,
                        "description": "Makes a retry safe: the same key with the same body returns the first result; with another body it is a 409.",
                        "schema": {
                            "type": "string",
                            "pattern": "^[A-Za-z0-9_-]{16,100}$"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmissionWrite"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Received",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubmissionResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Replayed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubmissionResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "410": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "A personal access token (arkon_pat_…). Scopes: read:pages, write:pages, read:posts, write:posts, write:categories, write:tags, read:media, write:media, read:forms, read:form_entries, read:users."
            }
        },
        "responses": {
            "Error": {
                "description": "An error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "fields": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "PaginationMeta": {
                "type": "object",
                "required": [
                    "page",
                    "per_page",
                    "total",
                    "total_pages"
                ],
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                }
            },
            "PaginationLinks": {
                "type": "object",
                "required": [
                    "self",
                    "first",
                    "prev",
                    "next",
                    "last"
                ],
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "next": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "last": {
                        "type": "string"
                    }
                }
            },
...
            "FormField": {
                "type": "object",
                "required": [
                    "id",
                    "type",
                    "label",
                    "required",
                    "choices"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    },
                    "placeholder": {
                        "type": "string"
                    },
                    "default_value": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "choices": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "required": [
                                "label",
                                "value"
                            ],
                            "properties": {
                                "label": {
                                    "type": "string"
                                },
                                "value": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "row": {
                        "type": "string"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "condition": {
                        "type": [
                            "object",
                            "null"
                        ]
                    }
                }
            },
            "Form": {
                "type": "object",
                "required": [
                    "id",
                    "version",
                    "name",
                    "description",
                    "submit_label",
                    "accepting_submissions",
                    "fields",
                    "submit_url"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "submit_label": {
                        "type": "string"
                    },
                    "accepting_submissions": {
                        "type": "boolean"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FormField"
                        }
                    },
                    "submit_url": {
                        "type": "string"
                    }
                }
            },
            "FormSummary": {
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "status",
                    "version",
                    "published_version",
                    "live_version",
                    "has_unpublished_changes",
                    "entries",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer"
                    },
                    "published_version": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "live_version": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "has_unpublished_changes": {
                        "type": "boolean"
                    },
                    "entries": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                }
            },
            "SubmissionWrite": {
                "type": "object",
                "required": [
                    "fields"
                ],
                "properties": {
                    "fields": {
                        "type": "object",
                        "description": "Values by field id."
                    }
                }
            },
            "SubmissionResult": {
                "type": "object",
                "required": [
                    "form",
                    "form_version",
                    "received",
                    "replayed",
                    "confirmation"
                ],
                "properties": {
                    "form": {
                        "type": "string"
                    },
                    "form_version": {
                        "type": "integer"
                    },
                    "received": {
                        "type": "boolean"
                    },
                    "replayed": {
                        "type": "boolean"
                    },
                    "confirmation": {
                        "type": "object",
                        "required": [
                            "type"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "message",
                                    "redirect"
                                ]
                            },
                            "message": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        }
    }
}